Fixes #20701: Fixing the unit test run error
