+++ tests/unit/plugins/to_validate/test_node_settings_load_parameters.py
@@ -1,6 +1,6 @@
+from __future__ import absolute_import, division, print_function
 import json
 import unittest
-from __future__ import absolute_import, division, print_function
 from unittest.mock import patch
 from plugins.modules import node_settings
 from ansible.module_utils import basic
