+++ tests/unit/plugins/to_validate/test_server_settings_load_parameters.py
@@ -1,6 +1,6 @@
+from __future__ import absolute_import, division, print_function
 import json
 import unittest
-from __future__ import absolute_import, division, print_function
 from unittest.mock import patch
 from ansible.module_utils import basic
 from plugins.modules import server_settings
